Implement working option generation solution
